Core: mm: Implement a simple frame_allocator to manage the frames
Implement the allocation and deallocation of available physical pages,
while providing interfaces for allocating and deallocating frames
(`frame_alloc` and `frame_dealloc`). In this context, the RAII
(Resource Acquisition Is Initialization) concept is employed to
encapsulate `PhysPageNum` as `FrameTracker`, allowing this tracker
to automatically manage the lifecycle of allocated physical page frames.

Signed-off-by: zq <zhouquan1511@163.com>