rCore: mm: Construction of Kernel/Application Address Spaces
The `MemorySet` is constructed based on [elf_data] from either the kernel
or the application. The corresponding methods are {new_kernel/from_elf}.
In the kernel, apart from the [trampoline] section, we establish identity
mappings for all other sections, for three reasons:

1) This allows us to access various segments of the kernel as we did before
   enabling the paging mechanism, without the need to adjust the kernel
   memory layout in "os/src/linker.ld".
2) Even after enabling the paging mechanism, the kernel can still read and
   write these physical page frames in a purely software manner.
3) The mappings in the vicinity of the instruction that switches the `satp`
   must maintain continuity between the two address spaces.

Finally, we provide a simple example `remap_test` to validate the correctness
of the kernel's address space mappings (considering only access permissions).

Signed-off-by: zq <zhouquan1511@163.com>