rCore: fs: Implementation of easy-fs [block_dev_trait]
Specific block device drivers are required to implement these
two methods: {read_block/write_block}, and block_cache will
call it to access block device.

For users of the `easy-fs` file system, it will be plugged
into the kernel in the future, and later we will implement
an `easy-Fs-fuse` application for testing.

Signed-off-by: zq <zhouquan1511@163.com>