rCore: fs: Implementation of [easy-fs] block_cache
A block cache and a global block cache manager are designed,
where the block cache corresponds to a block of the underlying
block device. All data structures of 'easy-fs' will first fall
into the cache layer and then be uniformly written back to the
block device by the cache layer.

Signed-off-by: zq <zhouquan1511@163.com>